swap_management: Add API for reclaiming memory of all processes
This adds the function ReclaimAllProcesses() to the API exposed
through D-Bus. The function reclaims memory of given types for
all processes, except for swap_management itself.

The new function is meant to be called just before the system
hibernates, with the goal of pushing most reclaimable memory
into zram, to reduce the size of the hibernation image.

BUG=b:301477158
TEST=tast run -var cycles=3 ${DUT} power.Hibernate
  => chrome://histograms/image.size is significantly smaller
     than without this change

Change-Id: I75bcaabad3ce56ca607a9fca1831fba387cc9797
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5124697
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 612b284e2f9e7d4606b93459843be6e204c4b18e