shill: tethering: finish previous SetTetheringEnabled async call when a new action comes
This change handles the case when a SetTetheringEnabled action is
ongoing, another SetTetheringEnabled action is issued before the
previous one is done.

When the new action is to abort a starting action, terminate the
previous action with result code abort and go through the stop process.
Return the stop action result when it is done. For other interrupting
actions, reject it and continue processing the previous action.

BUG=b:323251708
TEST=FEATURES=test emerge-{BOARD} shill
TEST=on device, abort tethering and check the abort result

Change-Id: I51f39f28fd9442a14acedf92a9e0c861a26f75ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5276994
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
NOKEYCHECK=True
GitOrigin-RevId: e084fcbaba226492951154e0193afbf13ab8bac5