biod: cryptohome: Refine FP migration protos
There are 2 changes in the CL. First, the ListLegacyRecordsReply proto
needs a status field just like other reply fields because we can't use
empty list to represent failure. Encountering error for this API and
having no legacy templates should lead to different actions.

The second change is that EnrollLegacyTemplate shouldn't be using the
OnStartEnrollSession callback because it doesn't create a session. It
should simply return a boolean status of whether it failed (of course we
can turn it into reply proto as well, but cryptohome usually doesn't
care about why it fails).

BUG=b:320541241
TEST=cryptohome/biod_proxy unittests

Change-Id: Ie383c9866aacbc9fd099b73a5c8a8cb034b6eafd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5353874
Reviewed-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Patryk Duda <patrykd@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Zi Lin <lziest@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 7c42bd3da69d65a1f5e01dc485ea9e3be78e1ae8