cryptohome: Add GetPinWeaverInfo API
This API can be used to check whether there is any PinWeaver credential
present on the device.

This is introduced to let hiberman check whether it can enable
hibernation (because hibernation is currently not compatible with
PinWeaver mechanism), but it can potentially be used for other scenarios
as well.

BUG=b:329020956
TEST=cryptohome unittests

Change-Id: I983dd762715f31c49d1c88506aea1eceb942f9cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5358898
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 5d0a2d778fa447548cb696db7436349f418ece40