power: Allow to specify if a suspend delay is applicable during key eviction
When key eviction is enabled the device key is evicted when the
system suspends. The key is only restored after user authentication
on resume. While the key is evicted some processes remain frozen to
prevent them from accessing cryptohome, which is unavailable during
key eviction. These processes might have registered a suspend delay
with powerd. Before suspending powerd waits for such processes to
indicate that they are ready to suspend. Waiting for a process that
is frozen would be futile. Add a field to RegisterSuspendDelayRequest
to indicate whether the suspend delay is applicable during key
eviction. If the field is not set assume that the delay is applicable
to maintain current behavior.

BUG=b:329139212
TEST=USE=key_eviction FEATURES=test P2_\
  TEST_FILTER="SuspendDelayControllerTest.*" \
  emerge-brya power_manager

Change-Id: Ib37ed06bf7db047770912f42ba5c11a69685ab72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5493312
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: c12333f6754e00198f1cf32119fe2a9c878edae0