shill: Refactor HotspotDevice creation to allow async response
Update the device creation flow in TetheringManager and WiFiProvider to
allow WiFiProvider to create the HotspotDevice asynchronously. This is
necessary for multi-interface Concurrency since WiFiProvider may be
required to perform slow operations like deleting existing interfaces
before it creates the HotspotDevice object.

The new flow is TetheringManager calls
WiFiProvider::RequestHotspotDeviceCreation, which triggers WiFiProvider
to create the HotspotDevice and send it to HotspotManager whenever it
is ready via TetheringManager::OnDeviceCreated. WiFiProvider makes no
guarantee on the sequencing of the its call to
TetheringManager::OnDeviceCreated, so TetheringManager must be prepared
for this to happen synchronously or asynchronously. Because
Multi-Interface Concurrency is not yet implemented, the asynchronous
call is always dispatched immediately when the request is made.

Some unit tests changes were required to enable this change, such as
introducing a new MockTetheringManager class.

BUG=b:328073379
TEST=FEATURES=test emerge-brya shill
TEST=Deploy and test tethering functionality. Can still enable tethering
and observe a working tethering device. Nearby device can see the
hotspot network in scans.

Change-Id: I8a65801a249f39702af1cbff90e152372b047f27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5368666
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 96ee60574606b6caf0beeb6630cb649e873992a2