system_api: Add Input & Output to PrepareAuthFactor
The goal is to get rid of the specialized GetRecoveryRequest operation
and instead fold it into the more normal AuthenticateAuthFactor process
by having it be done as part of a Prepare step. However, because the
GetRecoveryRequest operation takes input and produces output, this
requires adding some of the same ability to Prepare.

BUG=b:269619280
TEST=built protos & ran cryptohome unit tests

Change-Id: I57d965aaddad47dcfd1fee5b5523788758b95804
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5380864
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 86a6845a82c517e43d3a3cca03e73e00cd4eb7fe