shill: Make "priority" argument mandatory for P2P
Makes priority mandatory in accordance with the design in
go/cros-wifi-concurrency.

The value is still a stub, but it is now required in API calls.

The API is currently unused other than in tests (which are updated to
account for this in the dependee CL) so this is a safe transition.

BUG=b:328074924
TEST=FEATURES=test emerge-brya shill

Cq-Depend: chromium:5380610
Change-Id: Iecda6d2647e0a6220e887d8cadb2deb41be2b759
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5385900
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
NOKEYCHECK=True
GitOrigin-RevId: ce43473fa6b1f441a1d952f15993b37be06c82f7