cryptohome: Make Prepare work for recovery factor.
This finally adds in enough proto -> struct conversion for the inputs
needed to make the prepare operation for recovery work. This also means
we can convert the GetRecoveryRequest operation to just translate into a
PrepareAuthFactor request operation under the covers. Eventually we can
remove it once the usage of the old RPC is gone.

This also gets rid of the special member variable that AuthSession was
using to store the most recent ephemeral public key, and instead we look
it up by finding the active recovery token.

BUG=b:269619280
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=tast run $DUT cryptohome.Recovery*

Change-Id: I8cd1373edfcd50e0fabd383877cfde992867cb00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5403667
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
NOKEYCHECK=True
GitOrigin-RevId: cd33562b4a1ce36b133e3ab257dd2f470a7be859