modemfwd: Keep track of all modem's device ids seen
When a modem is detected and the modem matches the expected for the
variant, store the device id in /var/lib/modemfwd. Then use this
information to separate the metric kFailureReturnedByHelper into 2,
one for devices in which a modem was previously seen, and another for
devices in which a modem was never detected. Splitting the metrics
allows us to ignore metrics from SKUs which shipped without modems.

BUG=b:331839925
TEST=Added unit test. Verified the metrics reported on SKUs with modems
is different from the ones without modems.

Change-Id: I4469e252157839ab91d9812f38f02eb682128d69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5417081
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 551c2bc1ba42d0759318b2330e64be6abdb6d9cf