shill: add network id of attached Network in Service properties
This patch adds a "NetworkID" property to the DBus Service properties.

When the Service is idle, has not been selected by a Device, and has no
attached Network, the "NetworkID" property is equal to 0.

When the Service is active, has been selected by a Device, and has an
attached Network, the "NetworkID" property is equal to the network id of
its attached Network as assigned by patchpanel.

In addition, a property change event for "NetworkID" is emitted when the
Service attaches or detaches a Network.

BUG=b:336883268
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Id393bcb02178a19eb18f41d8efc1c0a4167f4a0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5528607
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 3e4dc09fca3f77cace1e44c2958c54aa684c2a4e