vm_tools: concierge: allow for mic permission calls for bruschetta
Other guests such a Crostini are able to use the built-in chromebook
microphone for their apps, and there is a desire from Bruschetta users
to also have access.

This patch introduces the ability for VmPermissionService to accept
calls for setting mic-share for Bruschetta.

BUG=b:315225198
TEST=compile chrome with relevant changes

Change-Id: I7094872ef95a8b5e6c0e2488c5df6c897650e240
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5593558
Tested-by: maciek swiech <drmasquatch@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: maciek swiech <drmasquatch@google.com>
NOKEYCHECK=True
GitOrigin-RevId: c5114ff95b6d00c0de6c9c7744f4e4ed1eee7516