shill: vpn: Read wg link status
Add the logic to execute `wg show wg0 dump` to read the link status
periodically, including the last handshake time, and rx and tx bytes for
each peer. Expose the information via the Provider property on Service
so that it can be accessed by Chrome, crosh, and included in the
service dump.

BUG=b:285971413
TEST=FEATURES="test" emerge-$BOARD shill
TEST=tast run $DUT network.WireguardCrosh network.*wireguard*
TEST=Checked that the WireGuard Service contains the expected properties
  in /usr/local/lib/flimflam/test/list-connected-service.

Change-Id: I5d464682f6e3b69ce822bbb4281d08302059bd63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5611630
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: a736cbd100d30d7bacfb0ae8d26794009aa68aa4