cryptohome: Add SetUserDataStorageEnabled method.
The method will be used in Chrome by SkyVault project that
needs to mark (empty) MyFiles and Downloads directories as
read-only to prevent creation of local user data files as
all the files should be stored in cloud in this case.

By default MyFiles and Downloads are mounted at the sing-in
as read-write, so this call will override the default once
Chrome detects SkyVault enterprise policy as it'll stay
read-only till next sign-in or till the policy is reset.

See http://go/skyvault-readonly for more info.

BUG=b:342564051
TEST=New unittests added

Change-Id: Idf1f1c4313278de26521182d762d664ee38533f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5638312
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Aida Zolic <aidazolic@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 4a7ce2f6fa8de1178642f1727c021d353f1b33fe