modemfwd: Add new modemfwd-helper api to check if recovery is needed
This change is validated with the next CL in the series, which uses this
API to prepare the recovery directory and required files. It was tested
on Brya+FM101 with an updated manifest, to ensure FM101's recovery can
occur from /tmp using the prepared files.

BUG=b:325679411
TEST=This change is validated with the above steps on Brya+FM101

Change-Id: If516585ea39c3490cf12cdf9938af76bb82f1eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5726844
Reviewed-by: Madhav <madhavadas@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
NOKEYCHECK=True
GitOrigin-RevId: c2637150767d43e98a13dbe51b6862e27414d927