shill: cellular: Add roaming_ip_type to APN configurations
Add the option to specify the IP type of an APN when roaming. This
allows for a carrier configuration to have different IP type values when
roaming and when not roaming.

BUG=b:359064950
TEST=Test roaming:
Used a US Cellular SIM card, which always roams.
-Added a roaming_ip_type to the US cellular APN.
Verify that the roaming ip type is used instead of the ip_type.
-Removed the roaming ip type from the APN, and verify the ip_type is
used.
Test no roaming:
Used a callbox SIM card, which does not roam.
-Added a roaming_ip_type value to the callbox APN, and verify that the
roaming_ip_type is not used.

Change-Id: I1cf9b85dae09fc23ebbf4883c5c2ed1f44b4085c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5789536
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
NOKEYCHECK=True
GitOrigin-RevId: e65ab850ab2a5e5d18e8fbec3a2d8945b30bdfa7