mtpd: Add `RequestCopyFileFromLocal` API
Add a `RequestCopyFileFromLocal` API that does not hold callers
because copying operation could take a long time (hours). Callers
use the API to request a copying and get the result via
`CopyFileFromLocalCompleted` signal.

Note only one file copy could be served. Even though callers
are not held for the copying request call, subsequent requests
would not go through until the current one finishes. This is
because `mtpd` uses `libmtp` that is not thread safe. Copying
needs to be done in the main thread. While the copying is going
on, the main thread is occupied and no dbus calls would be served.
This CL does not change this behavior.

BUG=b:388332345
TEST=Manual

Change-Id: I0a388c05050b6131f700b28661dd9914867162f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6251664
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: 4a460415b47a3cec5bc26525a35e3e5cf0402f5e