vm_tools: concierge: add xattr-based vm type identification for disk images
During the migration period from classic crostini to containerless
guests, it will be beneficial to have some 'ground truth' as to a vm's
(or in this case, vm's disk image's) state.

Given that we already use extended attributes for other disk
characteristics, we store the vm type as an xattr as well.
Unfortunately, duing the call to CreateDiskImage we do not currently
pass the vm type directly, so for the moment we will infer for the
containerless case by noting when the containerless image download flag
has been set. Since we only care to differentiate crostini and
containerless at the moment this will drastically reduce the amount of
other changes that need to be made.

CL also includes a few previously omitted `unlink` calls during error
cases.

BUG=b:407992760
TEST=compile, deploy, create vm, check xattr manually

Change-Id: I64b921290ebb2e3cf4ba6f4987df631025aebee2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6490094
Commit-Queue: maciek swiech <drmasquatch@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: maciek swiech <drmasquatch@google.com>
NOKEYCHECK=True
GitOrigin-RevId: eaa0356921045fe376e6c35b3e57e23cad798726