cryptohome: Add D-Bus method to generate recovery_id
Adds a new D-Bus method to generate a fresh recovery_id and store it in
a shadow directory. This makes the recovery_id accessible on the login
screen before the user is logged in.

This change ensures that the recovery container (seed, increment) are
only stored in the users encrypted directory, which is only accessible
during an authenticated user session.

BUG=b:420743396
TEST=cros_run_unit_tests --board=${BOARD} --packages
chromeos-base/cryptohome

Change-Id: Iffe2458f3ed552534788de60f5d82340771e5061
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6733104
Tested-by: Mohammed Abdon <mohammedabdon@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Mohammed Abdon <mohammedabdon@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
NOKEYCHECK=True
GitOrigin-RevId: ec64bc1c4115d06c1f25832133f8389ddf6c3f8b