system_api: add IsBruschettaAvailable dbus constant
Much like with crostini and pluginVM, we are sometimes interested in
querying if bruschetta is available. Make a new endpoint available.

The implementation lives in chrome code, but requires the constant to be
defined in system_api to be available to be called.

BUG=b:432480112
TEST=None

Change-Id: I44387f258be88940aba74ec6e26ab04ffde3e121
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6767170
Tested-by: maciek swiech <drmasquatch@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: maciek swiech <drmasquatch@google.com>
NOKEYCHECK=True
GitOrigin-RevId: 38881b9f6c005f4587b69f58c0bfa9fdb396e1c5