system_api: Add dbus constants for FjordOobeService
This endpoint allows for communication to OOBE to indicate that touch
controller setup is done and the next UI should be shown. The
implementation will live in the ChromeOS OOBE code.

Adding the constant here is required to be referenced from OOBE.

BUG=b:415052190
TEST=none

Change-Id: I715892b14a379b0d00d11f3cca15f8d79cc6e348
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6821049
Tested-by: Courtney Wong <courtneywong@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Courtney Wong <courtneywong@chromium.org>
NOKEYCHECK=True
GitOrigin-RevId: bcc50496a272e371384938f9be1e16fe09f0793c